Add canonical definition of SHA (from ocaml-git)
The SHA module in ocaml-git contains S, H, DIGEST and IO modules.
But to coerce any module to these canonicals definitions, we must flatten
S, H, DIGEST and IO outside of SHA or we must expect the exactly same
definition of SHA (with pretty, pp or compare or Map and Set specific
module of ocaml-git).

It's really not pleasant to use SHA of ocaml-git and I know what is the
cabal-hell, but this, it's really a functor/module hell for nothing.

Because we expect the exactly same definition of DIGEST, now we have the
useless dependencie cstruct in this project. If you have a better solution,
tell me!