Read passwords without lambda-term
The binaries that github-unix ships can read passwords, though the
library itself does not do it.
This makes the github-unix -> lambda-term dependency a bit heavy.

This factors out this feature in a small private library that is based
on the dune-release implementation:

https://github.com/ocamllabs/dune-release/blob/995bd1b3abc9ed54b1d9379f7046a23adb42e7d3/lib/stdext.ml#L37-L45

On windows, the unix calls are noops so this probably displays the
password.

Closes #237.